requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2 